requirements.txt
GitPython
flask
slackeventsapi
slackclient
zappa
